FLEX-5052 ~ Moves firmware history update functionality to domain adapter.
