+++ osgp/platform/osgp-adapter-domain-core/src/main/java/org/opensmartgridplatform/adapter/domain/core/application/services/FirmwareManagementService.java
@@ -8,9 +8,26 @@
 package org.opensmartgridplatform.adapter.domain.core.application.services;
 
 import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Date;
 import java.util.List;
+import java.util.Locale;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 import org.opensmartgridplatform.domain.core.entities.Device;
+import org.opensmartgridplatform.domain.core.entities.DeviceFirmwareFile;
+import org.opensmartgridplatform.domain.core.entities.DeviceModel;
+import org.opensmartgridplatform.domain.core.entities.FirmwareFile;
+import org.opensmartgridplatform.domain.core.entities.FirmwareModule;
+import org.opensmartgridplatform.domain.core.entities.Manufacturer;
+import org.opensmartgridplatform.domain.core.repositories.DeviceFirmwareFileRepository;
+import org.opensmartgridplatform.domain.core.repositories.DeviceModelRepository;
+import org.opensmartgridplatform.domain.core.repositories.DeviceRepository;
+import org.opensmartgridplatform.domain.core.repositories.FirmwareFileRepository;
+import org.opensmartgridplatform.domain.core.repositories.ManufacturerRepository;
+import org.opensmartgridplatform.domain.core.valueobjects.FirmwareModuleType;
 import org.opensmartgridplatform.domain.core.valueobjects.FirmwareUpdateMessageDataContainer;
 import org.opensmartgridplatform.domain.core.valueobjects.FirmwareVersion;
 import org.opensmartgridplatform.dto.valueobjects.FirmwareVersionDto;
@@ -25,6 +42,7 @@ import org.opensmartgridplatform.shared.infra.jms.ResponseMessageResultType;
 import org.opensmartgridplatform.shared.validation.Identification;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -33,6 +51,23 @@ import org.springframework.transaction.annotation.Transactional;
 public class FirmwareManagementService extends AbstractService {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(DeviceManagementService.class);
+
+    private static final String INSTALLER = "Inserted to match the version reported by the device.";
+
+    @Autowired
+    private DeviceFirmwareFileRepository deviceFirmwareFileRepository;
+
+    @Autowired
+    private DeviceModelRepository deviceModelRepository;
+
+    @Autowired
+    private DeviceRepository deviceRepository;
+
+    @Autowired
+    private FirmwareFileRepository firmwareFileRepository;
+
+    @Autowired
+    private ManufacturerRepository manufacturerRepository;
 
     /**
      * Constructor
@@ -100,6 +135,9 @@ public class FirmwareManagementService extends AbstractService {
         }
         final List<FirmwareVersion> firmwareVersions = this.domainCoreMapper.mapAsList(firmwareVersionsDto,
                 FirmwareVersion.class);
+
+        this.checkFirmwareHistory(ids.getOrganisationIdentification(), ids.getDeviceIdentification(), firmwareVersions);
+
         final ResponseMessage responseMessage = ResponseMessage.newResponseMessageBuilder()
                 .withIds(ids)
                 .withResult(result)
@@ -108,6 +146,126 @@ public class FirmwareManagementService extends AbstractService {
                 .withMessagePriority(messagePriority)
                 .build();
         this.webServiceResponseMessageSender.send(responseMessage);
+    }
+
+    private void checkFirmwareHistory(final String organisationIdentification, final String deviceId,
+            final List<org.opensmartgridplatform.domain.core.valueobjects.FirmwareVersion> firmwareVersions) {
+        final List<org.opensmartgridplatform.domain.core.valueobjects.FirmwareVersion> versionsNotInHistory = this
+                .checkFirmwareHistoryForVersion(organisationIdentification, deviceId, firmwareVersions);
+        for (final org.opensmartgridplatform.domain.core.valueobjects.FirmwareVersion firmwareVersion : versionsNotInHistory) {
+            LOGGER.info("Firmware version {} is not in history of device {}, we are trying to add it", firmwareVersion,
+                    deviceId);
+            this.tryToAddFirmwareVersionToHistory(organisationIdentification, deviceId, firmwareVersion);
+        }
+    }
+
+    /**
+     * @param organisationIdentification
+     *            the organisation the device we want to check belongs to
+     * @param deviceId
+     *            the id of the device we are checking
+     * @param firmwareVersions
+     *            the list of firmware versions to check if they are in the
+     *            history of the devices firmware history
+     * @return a list of firmware versions not present in the the devices
+     *         firmware history
+     * @throws FunctionalException
+     */
+    public List<FirmwareVersion> checkFirmwareHistoryForVersion(final String organisationIdentification,
+            final String deviceId, final List<FirmwareVersion> firmwareVersions) {
+
+        if (firmwareVersions.isEmpty()) {
+            return firmwareVersions;
+        }
+        // copy input parameter
+        final List<FirmwareVersion> firmwareVersionsToCheck = new ArrayList<>();
+        firmwareVersionsToCheck.addAll(firmwareVersions);
+
+        // get history
+        final Device device = this.deviceRepository.findByDeviceIdentification(deviceId);
+        final List<DeviceFirmwareFile> deviceFirmwareFiles = this.deviceFirmwareFileRepository
+                .findByDeviceOrderByInstallationDateAsc(device);
+        final List<FirmwareVersion> firmwareVersionsInHistory = deviceFirmwareFiles.stream()
+                .map(d -> d.getFirmwareFile().getModuleVersions().entrySet())
+                .flatMap(Collection::stream)
+                .map(e -> new FirmwareVersion(FirmwareModuleType.forDescription(e.getKey().getDescription()),
+                        e.getValue()))
+                .collect(Collectors.toList());
+
+        // remove the history versions
+        firmwareVersionsToCheck.removeAll(firmwareVersionsInHistory);
+
+        return firmwareVersionsToCheck;
+    }
+
+    public void tryToAddFirmwareVersionToHistory(final String organisationIdentification,
+            final String deviceIdentification, final FirmwareVersion firmwareVersion) {
+
+        final FirmwareModule module = createFirmwareModule(firmwareVersion);
+        final Device device = this.deviceRepository.findByDeviceIdentification(deviceIdentification);
+        final List<FirmwareFile> firmwareFiles = this
+                .getAvailableFirmwareFilesForDeviceModel(organisationIdentification, device.getDeviceModel());
+
+        // check each file for the module and the version as returned by the
+        // device
+        boolean recordAdded = false;
+
+        for (final FirmwareFile file : firmwareFiles) {
+            final Map<FirmwareModule, String> moduleVersions = file.getModuleVersions();
+            if (moduleVersions.containsKey(module) && moduleVersions.get(module).equals(firmwareVersion.getVersion())) {
+
+                // file found, insert a record into the history
+                final DeviceFirmwareFile deviceFirmwareFile = new DeviceFirmwareFile(device, file, new Date(),
+                        INSTALLER);
+                // this.saveDeviceFirmwareFile(deviceFirmwareFile);
+                this.deviceFirmwareFileRepository.save(deviceFirmwareFile);
+                LOGGER.info("Firmware version {} added to device {}", firmwareVersion.getVersion(),
+                        deviceIdentification);
+
+                // we only want to add one record in history
+                recordAdded = true;
+                break;
+            }
+        }
+
+        if (!recordAdded) {
+            LOGGER.warn("No firmware file record found for: {} for device: {}", firmwareVersion, deviceIdentification);
+        }
+
+    }
+
+    private static FirmwareModule createFirmwareModule(final FirmwareVersion firmwareVersion) {
+        final String description = firmwareVersion.getFirmwareModuleType().getDescription().toLowerCase(
+                Locale.getDefault());
+        return new FirmwareModule(description);
+    }
+
+    private List<FirmwareFile> getAvailableFirmwareFilesForDeviceModel(final String organisationIdentification,
+            final DeviceModel deviceModel) {
+        final Manufacturer manufacturer = deviceModel.getManufacturer();
+
+        return this.findAllFirmwareFiles(organisationIdentification, manufacturer.getCode(),
+                deviceModel.getModelCode());
+    }
+
+    private List<FirmwareFile> findAllFirmwareFiles(final String organisationIdentification, final String manufacturer,
+            final String modelCode) {
+        List<FirmwareFile> firmwareFiles = new ArrayList<>();
+        if (manufacturer != null) {
+            final Manufacturer databaseManufacturer = this.manufacturerRepository.findByCode(manufacturer);
+            final DeviceModel databaseDeviceModel = this.deviceModelRepository
+                    .findByManufacturerAndModelCode(databaseManufacturer, modelCode);
+            if (databaseDeviceModel != null) {
+                firmwareFiles = this.firmwareFileRepository.findByDeviceModel(databaseDeviceModel);
+            }
+        } else {
+            final List<DeviceModel> deviceModels = this.deviceModelRepository.findByModelCode(modelCode);
+            for (final DeviceModel deviceModel : deviceModels) {
+                firmwareFiles.addAll(this.firmwareFileRepository.findByDeviceModel(deviceModel));
+            }
+        }
+
+        return firmwareFiles;
     }
 
     // === SWITCH TO OTHER FIRMWARE VERSION ===
