+++ osgp/platform/osgp-adapter-ws-core/src/main/java/org/opensmartgridplatform/adapter/ws/core/endpoints/FirmwareManagementEndpoint.java
@@ -272,7 +272,6 @@ public class FirmwareManagementEndpoint {
                 organisationIdentification, deviceId);
 
         final GetFirmwareVersionResponse response = new GetFirmwareVersionResponse();
-        List<org.opensmartgridplatform.domain.core.valueobjects.FirmwareVersion> firmwareVersions = new ArrayList<>();
         try {
             final ResponseMessage message = this.firmwareManagementService
                     .dequeueGetFirmwareResponse(request.getAsyncRequest().getCorrelationUid());
@@ -280,7 +279,8 @@ public class FirmwareManagementEndpoint {
                 response.setResult(OsgpResultType.fromValue(message.getResult().getValue()));
                 if (message.getDataObject() != null) {
                     final List<FirmwareVersion> target = response.getFirmwareVersion();
-                    firmwareVersions = (List<org.opensmartgridplatform.domain.core.valueobjects.FirmwareVersion>) message
+                    @SuppressWarnings("unchecked")
+                    final List<org.opensmartgridplatform.domain.core.valueobjects.FirmwareVersion> firmwareVersions = (List<org.opensmartgridplatform.domain.core.valueobjects.FirmwareVersion>) message
                             .getDataObject();
                     target.addAll(this.firmwareManagementMapper.mapAsList(firmwareVersions, FirmwareVersion.class));
                 } else {
@@ -290,22 +290,8 @@ public class FirmwareManagementEndpoint {
         } catch (final Exception e) {
             this.handleException(e);
         }
-        this.checkFirmwareHistory(organisationIdentification, deviceId, firmwareVersions);
 
         return response;
-    }
-
-    private void checkFirmwareHistory(final String organisationIdentification, final String deviceId,
-            final List<org.opensmartgridplatform.domain.core.valueobjects.FirmwareVersion> firmwareVersions)
-            throws FunctionalException {
-        final List<org.opensmartgridplatform.domain.core.valueobjects.FirmwareVersion> versionsNotInHistory = this.firmwareManagementService
-                .checkFirmwareHistoryForVersion(organisationIdentification, deviceId, firmwareVersions);
-        for (final org.opensmartgridplatform.domain.core.valueobjects.FirmwareVersion firmwareVersion : versionsNotInHistory) {
-            LOGGER.info("Firmware version {} is not in history of device {}, we are trying to add it", firmwareVersion,
-                    deviceId);
-            this.firmwareManagementService.tryToAddFirmwareVersionToHistory(organisationIdentification, deviceId,
-                    firmwareVersion);
-        }
     }
 
     // === MANUFACTURERS LOGIC ===
